pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "allora-forge-builder-kit"
version = "0.1.0"
description = "Allora ML Workflow Utilities"
authors = [{ name = "Tim DeLise", email = "tim@alloralabs.xyz" }]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "pandas",
    "numpy",
    "scikit-learn",
    "xgboost",
    "scipy",
    "statsmodels",
    "optuna",
    "matplotlib",
    "ta",
    "cloudpickle",
    "python-dotenv",
    "PyYAML",
    "allora-sdk",
    "requests",
    "dill",
]

[project.urls]
Homepage = "https://github.com/allora-network/allora-forge-builder-kit"

[tool.setuptools.packages.find]
include = ["allora_forge_builder_kit"]